+++ src/main/java/chess/repository/exception/DataAccessException.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.repository.exception;
 
 public class DataAccessException extends RuntimeException {
     public DataAccessException(String message) {
