+++ src/main/java/chess/repository/utility/ResultSetMapper.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.repository.utility;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
