+++ src/test/java/chess/model/board/ChessBoardFixture.java
@@ -1,4 +1,4 @@
-package chess.model.board.util;
+package chess.model.board;
 
 import chess.model.piece.*;
