+++ src/test/java/chess/model/board/ChessBoardTest.java
@@ -4,8 +4,6 @@ import chess.model.piece.*;
 import chess.model.position.File;
 import chess.model.position.Position;
 import chess.model.position.Rank;
-import chess.model.board.util.ChessBoardFixture;
-import chess.model.board.util.TestChessBoardGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
