+++ src/test/java/chess/repository/dao/PieceDaoTest.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.repository.dao;
 
 import chess.model.board.ChessBoard;
 import chess.model.board.ChessBoardInitializer;
@@ -8,6 +8,7 @@ import chess.model.piece.Side;
 import chess.model.position.File;
 import chess.model.position.Position;
 import chess.model.position.Rank;
+import chess.repository.DataBaseCleaner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
